* chore(graphql): validate ID

Fix ARGOS-SERVER-ZEH

* chore(graphql): ignore unauthenticated errors

Fix ARGOS-SERVER-YBV

* chore: validate UserAccessToken name

Fix ARGOS-SERVER-ZDD

* chore: guard against invalid access token id

Fix ARGOS-SERVER-YVJ

* chore: guard for invalid account ids

Fix ARGOS-SERVER-YZ7

* chore: guard against invalid id

Fix ARGOS-SERVER-YVH

* chore: guard against invalid email

Fix ARGOS-SERVER-YQY

* chore: remove useless export

+++ apps/backend/src/graphql/definitions/UserAccessToken.ts
@@ -5,10 +5,11 @@ import { createCliAuthCode } from "@/auth/cli";
 import { UserAccessToken, UserAccessTokenScope } from "@/database/models";
 import { hashToken } from "@/database/services/crypto";
 import { transaction } from "@/database/transaction";
+import { isValidPgBigInt } from "@/database/util/biginteger";
 
 import type { IResolvers } from "../__generated__/resolver-types";
 import { getAccessibleAccounts } from "../services/account";
-import { badUserInput, forbidden } from "../util";
+import { badUserInput, forbidden, invalidId } from "../util";
 
 const { gql } = gqlTag;
 
@@ -182,6 +183,10 @@ export const resolvers: IResolvers = {
       }
 
       const { id } = args.input;
+
+      if (!isValidPgBigInt(id)) {
+        throw invalidId();
+      }
 
       await UserAccessToken.query().delete().where({
         id,
